[build-vm] Allow over committing memory on tmpfs
When the tmpfs worker instances combined size is more than the available
memory, the image fallocate in one go starts to fail randomly. This is due
to tmpf trying to fit the whole requested block in memory. Allocating it
in smaller chunks allows swapping out some of it as needed.

Most of the builds do not use anywhere near all of the disk space, so
it's safe to increase the size beyond memory limits without risking
excessive swapping.